fix: 显示-多屏拓展模式设置顺序调整 优化设置项顺序以及提供友好提示
Log: 显示-多屏拓展模式设置顺序调整
Task: https://pms.uniontech.com/task-view-212931.html
Influence: 调整主屏设置窗口的设置项顺序，并添加新文案
